requirements.txt
gunicorn
firebase_admin
flask
python-dotenv
pyjwt